+++ src/sniff_fns.js
@@ -4,7 +4,9 @@ var isChromeMobile = require('./is_browser_chrome_mobile');
 var isFF = require('./is_browser_firefox');
 var isOpera = require('./is_browser_opera');
 var isSafari = require('./is_browser_safari');
+var isIE = require('./is_browser_ie');
 var whichBrowser = require('./which_browser');
+var whichBrowserVersion = require('./which_browser_version');
 
 // OSes
 var isAndroid = require('./is_android');
@@ -28,7 +30,9 @@ var sniffFns = {
       isFirefox: isFF,
       isOpera: isOpera,
       isSafari: isSafari,
-      whichBrowser: whichBrowser
+      isIE: isIE,
+      whichBrowser: whichBrowser,
+      whichBrowserVersion: whichBrowserVersion
     },
     isIphone: isIphone,
     isIpad: isIpad,