+++ CHANGELOG.md
@@ -1,1 +1,7 @@
+## Unreleased
+
+- Add `isIE` function
+- Add `whichBrowserVersion` function which supports Chrome, Mobile Chrome,
+  Firefox, IE and Opera.
+
 ## v0.1.0 (19 Dec 2014)